Fix token fiat prices
