+++ apps/hyperdrive-trading/src/ui/landing/Landing.tsx
@@ -18,7 +18,7 @@ import { AssetStack } from "src/ui/markets/AssetStack";
 import { formatTermLength2 } from "src/ui/markets/formatTermLength";
 import { MARKET_DETAILS_ROUTE } from "src/ui/markets/routes";
 import { RewardsTooltip } from "src/ui/rewards/RewardsTooltip";
-import { useTokenFiatPrice } from "src/ui/token/hooks/useTokenFiatPrices";
+import { useTokenFiatPrice } from "src/ui/token/hooks/useTokenFiatPrice";
 import { useYieldSourceRate } from "src/ui/vaults/useYieldSourceRate";
 
 export function Landing(): ReactElement | null {
@@ -98,15 +98,16 @@ function PoolRow({ hyperdrive }: { hyperdrive: HyperdriveConfig }) {
     hyperdriveAddress: hyperdrive.address,
   });
   const isFiatPriceEnabled = !isTestnetChain(chainInfo.id);
-  const { fiatPrice } = useTokenFiatPrice({
-    tokenAddress: isFiatPriceEnabled
-      ? hyperdrive.poolConfig.baseToken
-      : undefined,
-  });
   const baseToken = findBaseToken({
     hyperdriveChainId: hyperdrive.chainId,
     hyperdriveAddress: hyperdrive.address,
     appConfig,
+  });
+  const { fiatPrice } = useTokenFiatPrice({
+    chainId: baseToken.chainId,
+    tokenAddress: isFiatPriceEnabled
+      ? hyperdrive.poolConfig.baseToken
+      : undefined,
   });
   let tvlLabel = `${formatCompact({
     value: presentValue || 0n,
